Starting 'locals' branch.
This is to implement local members. A local member is just a temporary storage for some expression result, which is evaluated and can be reused during the object value evaluation. Locals will replace the local fields inside imperative blocks.

An imperative block won't be a separate scope (LocalScope), while still be able to loop. Also, an imperative block won't produce a separate function in IR. Its evaluation will be performed inside a `claim` or or `proposition` one.