Starting 'no-dereference-by-default' branch.
A change to automatic link dereferencing rules. Links won't be dereferenced automatically as before. Instead, when searching the member of link, or converting it to other type, the link itself is tried first, if failed (the link has no such member, or can not be converted), then its target used.